todo controller geändert das jetzt das auch mit der auth middleware funktoniert

+++ controllers/taskController.js
@@ -1,8 +1,9 @@
 const Task = require('../models/task');
 
-exports.getAllTasks = async (req, res) => {
+exports.getTasks = async (req, res) => {
+    const userId = req.userId;
     try {
-        const allTodos = await Task.find(); // alle User
+        const allTodos = await Task.find({ userId }); // alle Todos zu einem User finden
         res.status(200).json(allTodos);
     }
     catch (err) {
@@ -48,6 +49,7 @@ exports.getTaskBySearch = async (req, res) => {
 };
 exports.createTask = async (req, res) => {
     const { text, tags, priority } = req.body;
+    const userId = req.userId;
 
     // Prüfen ob text da ist und nicht nur Leerzeichen ist
     if (!text || text.trim() === '') {
@@ -57,7 +59,8 @@ exports.createTask = async (req, res) => {
         const NewTodo = await Task.create({
             text: text.trim(),
             tags: tags || [],  // Falls nicht angegeben: leeres Array
-            priority: priority || 'Low'  // Falls nicht angegeben: Standard
+            priority: priority || 'Low',  // Falls nicht angegeben: Standard
+            userId: userId // die userId aus der middleware
         })
         res.status(201).json({ msg: "Neue Todo erstellt", NewTodo })
     }
@@ -68,13 +71,16 @@ exports.createTask = async (req, res) => {
 
 exports.delteTask = async (req, res) => {
     const id = req.params.id;
-
     try {
-        const delteTodo = await Task.findByIdAndDelete(id);
-        if (!delteTodo) { // wenn id nicht in den Todos gibt bzw wenn todo null ist
-            return res.status(404).json({ error: "Todo nicht gefunden" });
+        const task = await Task.findById(id);
+        if (!task) {
+            return res.status(404).json({ error: "Task nicht gefunden" });
         }
-        res.status(200).json({ msg: "erfolgreich gelöscht", delteTodo });
+        if (task.userId.toString() !== req.userId) { //in der Datenbank kein einfacher String ist, sondern ein BSON-Objekt ====> DESHALB toSring
+            return res.status(403).json({ error: "Du darfst diesen Task nicht löschen" });
+        }
+        const deltedTodo = await Task.findByIdAndDelete(id);
+        res.status(200).json({ msg: "erfolgreich gelöscht", deltedTodo });
     }
     catch (err) {
         res.status(500).json({ error: err.message });
@@ -83,11 +89,18 @@ exports.delteTask = async (req, res) => {
 
 exports.updateTask = async (req, res) => {
     const id = req.params.id;
-    const { text } = req.body;
+    const { text, } = req.body;
     if (!text || text.trim() === '') {
         return res.status(400).json({ error: 'Das Feld text darf nicht leer sein' });
     }
     try {
+        const task = await Task.findById(id);
+        if (!task) {
+            return res.status(404).json({ error: "Task nicht gefunden" });
+        }
+        if (task.userId.toString() !== req.userId) { //in der Datenbank kein einfacher String ist, sondern ein BSON-Objekt ====> DESHALB toSring
+            return res.status(403).json({ error: "Du darfst diesen Task nicht löschen" });
+        }
         const updatedTodo = await Task.findByIdAndUpdate(id, { text: text.trim() }, { new: true }); // new true damit die aktualisierte version zurückgegeben wird;
         if (!updatedTodo) {
             return res.status(404).json({ error: "Todo nicht gefunden" });
@@ -105,6 +118,13 @@ exports.patchTask = async (req, res) => {
     //    return res.status(400).json({ error: 'Das Feld text darf nicht leer sein' });
     //}
     try {
+        const task = await Task.findById(id);
+        if (!task) {
+            return res.status(404).json({ error: "Task nicht gefunden" });
+        }
+        if (task.userId.toString() !== req.userId) { //in der Datenbank kein einfacher String ist, sondern ein BSON-Objekt ====> DESHALB toSring
+            return res.status(403).json({ error: "Du darfst diesen Task nicht löschen" });
+        }
         const updatedTodo = await Task.findByIdAndUpdate(id, updates, { new: true }); // new true damit die aktualisierte version zurückgegeben wird;
         if (!updatedTodo) {
             return res.status(404).json({ error: "Todo nicht gefunden" });
